T0 基线快照: 8936c01 + VL 在途 22 文件 + 未跟踪必需件 (third_party/quickstart/backport/taskbook/wt_dump_ops) + test_assets 同步
VL 在途快照自 /4090disk2/GEHTP 工作树 (未评审原样落档): 多输入注入+binary 扩展+transpose perm 兜底+路线B run 支持
路径 shim: /disk2/{GEHTP,QCtools,Qwen35dev} → /4090disk2/* (104 视角路径 verbatim 兼容)

+++ compiler/include/hnnx/ir/op_extra.hpp
@@ -56,7 +56,8 @@ struct ExtraStridedSlice {
     uint32_t end_mask;       // +8
     uint32_t new_axes_mask;  // +12
     uint32_t shrink_axes;    // +16
-    uint32_t ranges_offset;  // +20 (const 池偏移, 指向 rank×3×u32 数组)
+    uint64_t ranges_offset;  // +24 (const 池偏移, 指向 rank×3×u32 数组; 64 位——
+                             //  16GB 池末层 ranges coff>4GB, 32 位截断读垃圾)
 };
 
 // Reshape: 目标形状(dims 数组直接进 extra, rank ≤ 5)
